Moved telekom code to designated dns_provider folder
